+++ src/data/input_output/data_source.py
@@ -7,7 +7,7 @@ from functools import cached_property
 from logging import Logger
 from typing import override, ClassVar
 
-from common import format_timestamp_date_time
+from common import format_timestamp_date_time, SharlyChessException
 from common.i18n import _
 from common.logger import get_logger
 from common.network import NetworkMonitor
@@ -205,6 +205,12 @@ class DataSource(IdentifiableEntity, ABC):
     def on_app_init(self):
         """Function to execute at the start of the server to initialize the data source."""
 
+    @property
+    def info_or_warning_message(self) -> tuple[str, bool]:
+        """Message displayed on the players update modal and the player search.
+        If the bool is set to True, show the message as a warning."""
+        return '', False
+
     # --------------------------------------------------------------------------
     # Players update
     # --------------------------------------------------------------------------
@@ -213,12 +219,6 @@ class DataSource(IdentifiableEntity, ABC):
     @abstractmethod
     def player_updater_fields(self) -> list[PlayerUpdaterField]:
         """Returns the player fields that can be updated by the data source."""
-
-    @property
-    def players_update_message(self) -> tuple[str, bool]:
-        """Message displayed on the players update modal.
-        If the bool is set to True, show the message as a warning."""
-        return '', False
 
     @abstractmethod
     async def get_player_matches(
@@ -278,6 +278,11 @@ class DataSource(IdentifiableEntity, ABC):
     def player_search_result_template(self) -> str:
         """Template containing the info to display for a player as a search result.
         Template takes [player] as a template variable."""
+
+    @property
+    @abstractmethod
+    def search_error_icon(self) -> str:
+        """Icon to display in the search results in case of an error."""
 
     @abstractmethod
     async def search_player(
@@ -347,17 +352,20 @@ class LocalDataSource(DataSource, ABC):
     def local_database_type(self) -> type[LocalSourceDatabase]:
         """The type of the local database used for this source."""
 
+    @cached_property
+    def database(self) -> LocalSourceDatabase:
+        return self.local_database_type()
+
     @property
-    def players_update_message(self) -> tuple[str, bool]:
-        database = self.local_database_type()
+    def info_or_warning_message(self) -> tuple[str, bool]:
         message = (
             _('Last update: {updated_at} (outdated)')
-            if database.is_outdated
+            if self.database.is_outdated
             else _('Last update: {updated_at}')
         )
         return (
-            message.format(updated_at=database.updated_at_str),
-            database.is_outdated,
+            message.format(updated_at=self.database.updated_at_str),
+            self.database.is_outdated,
         )
 
     @property
@@ -365,11 +373,22 @@ class LocalDataSource(DataSource, ABC):
         return self.local_database_type.file_path().exists()
 
     def on_app_init(self):
-        self.local_database_type().check()
+        self.database.check()
+
+    @property
+    def search_error_icon(self) -> str:
+        return 'bi-database-fill-dash'
 
     async def search_player(
         self, string: str, federation: str, page: int = 0, limit: int | None = None
     ) -> list[StoredPlayer]:
+        if not self.is_available:
+            raise SharlyChessException(
+                _(
+                    'This database is not installed '
+                    '(to install it: Menu > Data sources).'
+                )
+            )
         with self.local_database_type() as database:
             return database.search_player(string, federation, page, limit)
 
@@ -397,16 +416,44 @@ class OnlineDataSource(DataSource, ABC):
     @classmethod
     async def reload_connection_status(cls):
         cls._connection_last_checked_at = time.time()
-        if NetworkMonitor.connected():
-            cls.connection_status = await cls.check_connection()
+        if not NetworkMonitor.connected():
+            cls.connection_status = None
+        cls.connection_status = await cls.check_connection()
 
     @property
     def is_available(self) -> bool:
-        return NetworkMonitor.connected() and bool(self.connection_status)
+        return True
 
     @property
     def connection_last_checked_at_str(self) -> str:
         return format_timestamp_date_time(self._connection_last_checked_at)
+
+    @property
+    def search_error_icon(self) -> str:
+        return 'bi-cloud-fill-dash'
+
+    async def search_player(
+        self, string: str, federation: str, page: int = 0, limit: int | None = None
+    ) -> list[StoredPlayer]:
+        cls = self.__class__
+        cls._connection_last_checked_at = time.time()
+        if not NetworkMonitor.connected():
+            cls.connection_status = None
+            raise SharlyChessException(_('Not connected to internet'))
+        try:
+            players = await self._search_player(string, federation, page, limit)
+            cls.connection_status = True
+            return players
+        except SharlyChessException as exception:
+            cls.connection_status = False
+            raise exception
+
+    @abstractmethod
+    async def _search_player(
+        self, string: str, federation: str, page: int = 0, limit: int | None = None
+    ) -> list[StoredPlayer]:
+        """Search a player in the data source from a string.
+        Returns maximum *limit* results (no limit if *limit* is None)."""
 
 
 class FideDataSource(LocalDataSource):